requirements.txt
asgiref==3.10.0
certifi==2025.10.5
charset-normalizer==3.4.4
django==5.2.8
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
dj-database-url==2.2.0
gunicorn==22.0.0
httpx==0.28.1
idna==3.11
psycopg[binary]==3.2.12
python-dotenv==1.2.1
requests==2.32.5
resend==2.19.0
sentry-sdk==1.45.1
sqlparse==0.5.3
whitenoise==6.7.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
